Avoiding the double-reading problem of Net::HTTP
When issuing a request with Net::HTTP, you can provide a block in which you may read the
body chunk by chunk. However, it will continue to read the remainder fo the request body
afterwards, and so you're left needing to do all of your chunked work within the context
of the block.

It's a little ugly, but the block just has to be propagated forwards quite a ways.