Raise ArgumentError if required options are falsey
I was occasionally seeing "no implicit conversion of nil into String
(TypeError)" when calling buckets.find. This was happening all the way
in the `digest` call in the Signature class because the
`secret_access_key` had been set to `nil`. Since this value is set at
initialization time, we can be proactive with a helpful error message
when this happens.

Fixes #107