Issue 125: Object#content now aware of @content initilization
More details in #125:

With fc4d193 a change was made to S3::Object#initialize which
initializes the value of `@content` to `nil`. In Ruby this is enough to
satisfy `Object#defined?` (as `instance-variable`). This had an
unintended consequence of the `content(reload = false)` method always
return `nil` unless `reload = true`.

This commit will make the method more aware and address the issue in the
existing test suite.